+++ lib/experimental/interactivity-api/directive-processing.php
@@ -13,19 +13,21 @@
  * parent is null.
  *
  * @param array $parsed_block The parsed block.
- * @param array $source_block The source block.
- * @param array $parent_block The parent block.
  *
  * @return array The parsed block.
  */
-function gutenberg_interactivity_mark_root_blocks( $parsed_block, $source_block, $parent_block ) {
-	if ( ! isset( $parent_block ) && ! WP_Directive_Processor::has_root_block() ) {
-		WP_Directive_Processor::mark_root_block( $parsed_block );
+function gutenberg_interactivity_mark_root_interactive_blocks( $parsed_block ) {
+	if ( ! WP_Directive_Processor::has_interactive_root_block() ) {
+		$block_type     = WP_Block_Type_Registry::get_instance()->get_registered( $parsed_block['blockName'] );
+		$is_interactive = isset( $block_type->supports['interactivity'] ) && $block_type->supports['interactivity'];
+		if ( $is_interactive ) {
+			WP_Directive_Processor::mark_interactive_root_block( $parsed_block );
+		}
 	}
 
 	return $parsed_block;
 }
-add_filter( 'render_block_data', 'gutenberg_interactivity_mark_root_blocks', 10, 3 );
+add_filter( 'render_block_data', 'gutenberg_interactivity_mark_root_interactive_blocks', 10, 1 );
 
 /**
  * Processes the directives in the root blocks.
@@ -36,145 +38,16 @@ add_filter( 'render_block_data', 'gutenberg_interactivity_mark_root_blocks', 10,
  * @return string Filtered block content.
  */
 function gutenberg_process_directives_in_root_blocks( $block_content, $block ) {
-	if ( WP_Directive_Processor::is_marked_as_root_block( $block ) ) {
-		WP_Directive_Processor::unmark_root_block();
-
-		// Parse our own block delimiters for interactive and non-interactive blocks.
-		$parsed_blocks     = parse_blocks( $block_content );
-		$context           = new WP_Directive_Context();
-		$processed_content = '';
-		$namespace_stack   = array();
-
-		foreach ( $parsed_blocks as $parsed_block ) {
-			if ( 'core/interactivity-wrapper' === $parsed_block['blockName'] ) {
-				$processed_content .= gutenberg_process_interactive_block( $parsed_block, $context, $namespace_stack );
-			} elseif ( 'core/non-interactivity-wrapper' === $parsed_block['blockName'] ) {
-				$processed_content .= gutenberg_process_non_interactive_block( $parsed_block, $context, $namespace_stack );
-			} else {
-				$processed_content .= $parsed_block['innerHTML'];
-			}
-		}
-		return $processed_content;
+	if ( WP_Directive_Processor::is_marked_as_interactive_root_block( $block ) ) {
+		WP_Directive_Processor::unmark_interactive_root_block();
+		$context         = new WP_Directive_Context();
+		$namespace_stack = array();
+		return gutenberg_process_interactive_html( $block_content, $context, $namespace_stack );
 	}
 
 	return $block_content;
 }
-add_filter( 'render_block', 'gutenberg_process_directives_in_root_blocks', 20, 2 );
-
-/**
- * Marks the block as a children of an interactive block.
- *
- * @param array    $parsed_block The parsed block.
- * @param array    $source_block The source block.
- * @param WP_Block $parent_block The parent block.
- */
-function gutenberg_mark_chidren_of_interactive_block( $parsed_block, $source_block, $parent_block ) {
-	if (
-		isset( $parent_block ) &&
-		isset( $parent_block->block_type->supports['interactivity'] ) &&
-		$parent_block->block_type->supports['interactivity']
-	) {
-		WP_Directive_Processor::mark_children_of_interactive_block( $source_block );
-	}
-	return $parsed_block;
-}
-add_filter( 'render_block_data', 'gutenberg_mark_chidren_of_interactive_block', 100, 3 );
-
-/**
- * Adds a comment delimiter to mark if the block is interactive or not.
- *
- * @param string   $block_content The block content.
- * @param array    $block The full block, including name and attributes.
- * @param WP_Block $block_instance The block instance.
- */
-function gutenberg_mark_block_interactivity( $block_content, $block, $block_instance ) {
-	if (
-		isset( $block_instance->block_type->supports['interactivity'] ) &&
-		$block_instance->block_type->supports['interactivity']
-	) {
-		// Wraps the interactive block with a comment delimiter to be able to
-		// process it later.
-		return get_comment_delimited_block_content(
-			'core/interactivity-wrapper',
-			array(),
-			$block_content
-		);
-	} elseif ( WP_Directive_Processor::is_marked_as_children_of_interactive_block( $block ) ) {
-		// Wraps the non-interactive block with a comment delimiter to be able to
-		// skip it later.
-		return get_comment_delimited_block_content(
-			'core/non-interactivity-wrapper',
-			array(),
-			$block_content
-		);
-	}
-	return $block_content;
-}
-add_filter( 'render_block', 'gutenberg_mark_block_interactivity', 10, 3 );
-
-/**
- * Traverses the HTML of an interactive block, searching for Interactivity API
- * directives and processing them. For the inner blocks, it calls the
- * corresponding function depending on the wrapper type.
- *
- * @param array                $interactive_block The interactive block to process.
- * @param WP_Directive_Context $context The context to use when processing.
- * @param array                $namespace_stack Stack of namespackes passed by reference.
- *
- * @return string The processed HTML.
- */
-function gutenberg_process_interactive_block( $interactive_block, $context, &$namespace_stack ) {
-	$block_index              = 0;
-	$content                  = '';
-	$interactive_inner_blocks = array();
-
-	foreach ( $interactive_block['innerContent'] as $inner_content ) {
-		if ( is_string( $inner_content ) ) {
-			$content .= $inner_content;
-		} else {
-			// This is an inner block. It may be an interactive block or a
-			// non-interactive block.
-			$content                   .= '<wp-inner-blocks-' . $block_index . '></wp-inner-blocks-' . $block_index . '>';
-			$interactive_inner_blocks[] = $interactive_block['innerBlocks'][ $block_index++ ];
-		}
-	}
-
-	return gutenberg_process_interactive_html( $content, $context, $interactive_inner_blocks, $namespace_stack );
-}
-
-/**
- * Returns the HTML of a non-interactive block without processing the
- * directives. For the inner blocks, it calls the corresponding function
- * depending on the wrapper type.
- *
- * @param array                $non_interactive_block The non-interactive block to process.
- * @param WP_Directive_Context $context The context to use when processing.
- * @param array                $namespace_stack Stack of namespackes passed by reference.
- *
- * @return string The processed HTML.
- */
-function gutenberg_process_non_interactive_block( $non_interactive_block, $context, &$namespace_stack ) {
-	$block_index = 0;
-	$content     = '';
-	foreach ( $non_interactive_block['innerContent'] as $inner_content ) {
-		if ( is_string( $inner_content ) ) {
-			// This content belongs to a non interactive block and therefore it cannot
-			// contain directives. We add the HTML directly to the final output.
-			$content .= $inner_content;
-		} else {
-			// This is an inner block. It may be an interactive block or a
-			// non-interactive block.
-			$inner_block = $non_interactive_block['innerBlocks'][ $block_index++ ];
-
-			if ( 'core/interactivity-wrapper' === $inner_block['blockName'] ) {
-				$content .= gutenberg_process_interactive_block( $inner_block, $context, $namespace_stack );
-			} elseif ( 'core/non-interactivity-wrapper' === $inner_block['blockName'] ) {
-				$content .= gutenberg_process_non_interactive_block( $inner_block, $context, $namespace_stack );
-			}
-		}
-	}
-	return $content;
-}
+add_filter( 'render_block', 'gutenberg_process_directives_in_root_blocks', 10, 2 );
 
 /**
  * Processes interactive HTML by applying directives to the HTML tags.
@@ -191,7 +64,7 @@ function gutenberg_process_non_interactive_block( $non_interactive_block, $conte
  *
  * @return string The processed HTML.
  */
-function gutenberg_process_interactive_html( $html, $context, $inner_blocks = array(), &$namespace_stack = array() ) {
+function gutenberg_process_interactive_html( $html, $context, &$namespace_stack = array() ) {
 	static $directives = array(
 		'data-wp-interactive' => 'gutenberg_interactivity_process_wp_interactive',
 		'data-wp-context'     => 'gutenberg_interactivity_process_wp_context',
@@ -201,22 +74,12 @@ function gutenberg_process_interactive_html( $html, $context, $inner_blocks = ar
 		'data-wp-text'        => 'gutenberg_interactivity_process_wp_text',
 	);
 
-	$tags                   = new WP_Directive_Processor( $html );
-	$prefix                 = 'data-wp-';
-	$tag_stack              = array();
-	$inner_processed_blocks = array();
-	$inner_blocks_index     = 0;
+	$tags      = new WP_Directive_Processor( $html );
+	$prefix    = 'data-wp-';
+	$tag_stack = array();
 	while ( $tags->next_tag( array( 'tag_closers' => 'visit' ) ) ) {
 		$tag_name = $tags->get_tag();
 
-		// Processes the inner blocks.
-		if ( str_contains( $tag_name, 'WP-INNER-BLOCKS' ) && ! empty( $inner_blocks ) && ! $tags->is_tag_closer() ) {
-			if ( 'core/interactivity-wrapper' === $inner_blocks[ $inner_blocks_index ]['blockName'] ) {
-				$inner_processed_blocks[ strtolower( $tag_name ) ] = gutenberg_process_interactive_block( $inner_blocks[ $inner_blocks_index++ ], $context, $namespace_stack );
-			} elseif ( 'core/non-interactivity-wrapper' === $inner_blocks[ $inner_blocks_index ]['blockName'] ) {
-				$inner_processed_blocks[ strtolower( $tag_name ) ] = gutenberg_process_non_interactive_block( $inner_blocks[ $inner_blocks_index++ ], $context, $namespace_stack );
-			}
-		}
 		if ( $tags->is_tag_closer() ) {
 			if ( 0 === count( $tag_stack ) ) {
 				continue;
@@ -287,19 +150,7 @@ function gutenberg_process_interactive_html( $html, $context, $inner_blocks = ar
 		}
 	}
 
-	$processed_html = $tags->get_updated_html();
-
-	// Replaces the inner block tags with the content of each inner block
-	// processed.
-	if ( ! empty( $inner_processed_blocks ) ) {
-		foreach ( $inner_processed_blocks as $inner_block_tag => $inner_block_content ) {
-			if ( str_contains( $processed_html, $inner_block_tag ) ) {
-				$processed_html = str_replace( '<' . $inner_block_tag . '></' . $inner_block_tag . '>', $inner_block_content, $processed_html );
-			}
-		}
-	}
-
-	return $processed_html;
+	return $tags->get_updated_html();
 }
 
 /**