+++ packages/block-editor/src/hooks/content-lock-ui.js
@@ -4,40 +4,23 @@
 import { ToolbarButton, MenuItem } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
-import { useEffect, useRef, useCallback } from '@wordpress/element';
+import { useCallback } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import { store as blockEditorStore } from '../store';
 import { BlockControls, BlockSettingsMenuControls } from '../components';
+import { unlock } from '../lock-unlock';
 
-function StopEditingAsBlocksOnOutsideSelect( {
-	clientId,
-	stopEditingAsBlock,
-} ) {
-	const isBlockOrDescendantSelected = useSelect(
-		( select ) => {
-			const { isBlockSelected, hasSelectedInnerBlock } =
-				select( blockEditorStore );
-			return (
-				isBlockSelected( clientId ) ||
-				hasSelectedInnerBlock( clientId, true )
-			);
-		},
-		[ clientId ]
-	);
-	useEffect( () => {
-		if ( ! isBlockOrDescendantSelected ) {
-			stopEditingAsBlock();
-		}
-	}, [ isBlockOrDescendantSelected, stopEditingAsBlock ] );
-	return null;
-}
+// The implementation of content locking is mainly in this file, although the mechanism
+// to stop temporarily editing as blocks when an outside block is selected is on component StopEditingAsBlocksOnOutsideSelect
+// at block-editor/src/components/block-list/index.js.
+// Besides the components on this file and the file referenced above the implementation
+// also includes artifacts on the store (actions, reducers, and selector).
 
 function ContentLockControlsPure( { clientId, isSelected } ) {
 	const { getBlockListSettings, getSettings } = useSelect( blockEditorStore );
-	const focusModeToRevert = useRef();
 	const { templateLock, isLockedByParent, isEditingAsBlocks } = useSelect(
 		( select ) => {
 			const {
@@ -61,31 +44,15 @@ function ContentLockControlsPure( { clientId, isSelected } ) {
 		updateBlockListSettings,
 		__unstableSetTemporarilyEditingAsBlocks,
 	} = useDispatch( blockEditorStore );
+	const { stopEditingAsBlocks } = unlock( useDispatch( blockEditorStore ) );
 	const isContentLocked =
 		! isLockedByParent && templateLock === 'contentOnly';
 	const { __unstableMarkNextChangeAsNotPersistent, updateBlockAttributes } =
 		useDispatch( blockEditorStore );
 
-	const stopEditingAsBlock = useCallback( () => {
-		__unstableMarkNextChangeAsNotPersistent();
-		updateBlockAttributes( clientId, {
-			templateLock: 'contentOnly',
-		} );
-		updateBlockListSettings( clientId, {
-			...getBlockListSettings( clientId ),
-			templateLock: 'contentOnly',
-		} );
-		updateSettings( { focusMode: focusModeToRevert.current } );
-		__unstableSetTemporarilyEditingAsBlocks();
-	}, [
-		clientId,
-		updateSettings,
-		updateBlockListSettings,
-		getBlockListSettings,
-		__unstableMarkNextChangeAsNotPersistent,
-		updateBlockAttributes,
-		__unstableSetTemporarilyEditingAsBlocks,
-	] );
+	const stopEditingAsBlockCallback = useCallback( () => {
+		stopEditingAsBlocks( clientId );
+	}, [ clientId, stopEditingAsBlocks ] );
 
 	if ( ! isContentLocked && ! isEditingAsBlocks ) {
 		return null;
@@ -99,16 +66,8 @@ function ContentLockControlsPure( { clientId, isSelected } ) {
 		<>
 			{ showStopEditingAsBlocks && (
 				<>
-					<StopEditingAsBlocksOnOutsideSelect
-						clientId={ clientId }
-						stopEditingAsBlock={ stopEditingAsBlock }
-					/>
 					<BlockControls group="other">
-						<ToolbarButton
-							onClick={ () => {
-								stopEditingAsBlock();
-							} }
-						>
+						<ToolbarButton onClick={ stopEditingAsBlockCallback }>
 							{ __( 'Done' ) }
 						</ToolbarButton>
 					</BlockControls>
@@ -127,11 +86,12 @@ function ContentLockControlsPure( { clientId, isSelected } ) {
 									...getBlockListSettings( clientId ),
 									templateLock: false,
 								} );
-								focusModeToRevert.current =
+								const focusModeToRevert =
 									getSettings().focusMode;
 								updateSettings( { focusMode: true } );
 								__unstableSetTemporarilyEditingAsBlocks(
-									clientId
+									clientId,
+									focusModeToRevert
 								);
 								onClose();
 							} }