+++ packages/interactivity/src/directives.js
@@ -1,23 +1,15 @@
 /**
  * External dependencies
  */
-import {
-	useContext,
-	useMemo,
-	useEffect,
-	useRef,
-	useLayoutEffect,
-} from 'preact/hooks';
+import { useContext, useMemo, useRef } from 'preact/hooks';
 import { deepSignal, peek } from 'deepsignal';
 
 /**
  * Internal dependencies
  */
 import { createPortal } from './portals';
-import { useSignalEffect } from './utils';
-import { directive } from './hooks';
-import { SlotProvider, Slot, Fill } from './slots';
-import { navigate } from './router';
+import { useWatch, useInit } from './utils';
+import { directive, getScope, getEvaluate } from './hooks';
 
 const isObject = ( item ) =>
 	item && typeof item === 'object' && ! Array.isArray( item );
@@ -35,6 +27,64 @@ const mergeDeepSignals = ( target, source, overwrite ) => {
 		}
 	}
 };
+
+const newRule =
+	/(?:([\u0080-\uFFFF\w-%@]+) *:? *([^{;]+?);|([^;}{]*?) *{)|(}\s*)/g;
+const ruleClean = /\/\*[^]*?\*\/|  +/g;
+const ruleNewline = /\n+/g;
+const empty = ' ';
+
+/**
+ * Convert a css style string into a object.
+ *
+ * Made by Cristian Bote (@cristianbote) for Goober.
+ * https://unpkg.com/browse/goober@2.1.13/src/core/astish.js
+ *
+ * @param {string} val CSS string.
+ * @return {Object} CSS object.
+ */
+const cssStringToObject = ( val ) => {
+	const tree = [ {} ];
+	let block, left;
+
+	while ( ( block = newRule.exec( val.replace( ruleClean, '' ) ) ) ) {
+		if ( block[ 4 ] ) {
+			tree.shift();
+		} else if ( block[ 3 ] ) {
+			left = block[ 3 ].replace( ruleNewline, empty ).trim();
+			tree.unshift( ( tree[ 0 ][ left ] = tree[ 0 ][ left ] || {} ) );
+		} else {
+			tree[ 0 ][ block[ 1 ] ] = block[ 2 ]
+				.replace( ruleNewline, empty )
+				.trim();
+		}
+	}
+
+	return tree[ 0 ];
+};
+
+/**
+ * Creates a directive that adds an event listener to the global window or
+ * document object.
+ *
+ * @param {string} type 'window' or 'document'
+ * @return {void}
+ */
+const getGlobalEventDirective =
+	( type ) =>
+	( { directives, evaluate } ) => {
+		directives[ `on-${ type }` ]
+			.filter( ( { suffix } ) => suffix !== 'default' )
+			.forEach( ( entry ) => {
+				useInit( () => {
+					const cb = ( event ) => evaluate( entry, event );
+					const globalVar = type === 'window' ? window : document;
+					globalVar.addEventListener( entry.suffix, cb );
+					return () =>
+						globalVar.removeEventListener( entry.suffix, cb );
+				}, [] );
+			} );
+	};
 
 export default () => {
 	// data-wp-context
@@ -75,25 +125,33 @@ export default () => {
 	// data-wp-watch--[name]
 	directive( 'watch', ( { directives: { watch }, evaluate } ) => {
 		watch.forEach( ( entry ) => {
-			useSignalEffect( () => evaluate( entry ) );
+			useWatch( () => evaluate( entry ) );
 		} );
 	} );
 
 	// data-wp-init--[name]
 	directive( 'init', ( { directives: { init }, evaluate } ) => {
 		init.forEach( ( entry ) => {
-			useEffect( () => evaluate( entry ), [] );
+			// TODO: Replace with useEffect to prevent unneeded scopes.
+			useInit( () => evaluate( entry ) );
 		} );
 	} );
 
 	// data-wp-on--[event]
 	directive( 'on', ( { directives: { on }, element, evaluate } ) => {
-		on.forEach( ( entry ) => {
-			element.props[ `on${ entry.suffix }` ] = ( event ) => {
-				evaluate( entry, event );
-			};
-		} );
+		on.filter( ( { suffix } ) => suffix !== 'default' ).forEach(
+			( entry ) => {
+				element.props[ `on${ entry.suffix }` ] = ( event ) => {
+					evaluate( entry, event );
+				};
+			}
+		);
 	} );
+
+	// data-wp-on-window--[event]
+	directive( 'on-window', getGlobalEventDirective( 'window' ) );
+	// data-wp-on-document--[event]
+	directive( 'on-document', getGlobalEventDirective( 'document' ) );
 
 	// data-wp-class--[classname]
 	directive(
@@ -118,7 +176,7 @@ export default () => {
 							? `${ currentClass } ${ name }`
 							: name;
 
-					useEffect( () => {
+					useInit( () => {
 						// This seems necessary because Preact doesn't change the class
 						// names on the hydration, so we have to do it manually. It doesn't
 						// need deps because it only needs to do it the first time.
@@ -127,45 +185,10 @@ export default () => {
 						} else {
 							element.ref.current.classList.add( name );
 						}
-					}, [] );
+					} );
 				} );
 		}
 	);
-
-	const newRule =
-		/(?:([\u0080-\uFFFF\w-%@]+) *:? *([^{;]+?);|([^;}{]*?) *{)|(}\s*)/g;
-	const ruleClean = /\/\*[^]*?\*\/|  +/g;
-	const ruleNewline = /\n+/g;
-	const empty = ' ';
-
-	/**
-	 * Convert a css style string into a object.
-	 *
-	 * Made by Cristian Bote (@cristianbote) for Goober.
-	 * https://unpkg.com/browse/goober@2.1.13/src/core/astish.js
-	 *
-	 * @param {string} val CSS string.
-	 * @return {Object} CSS object.
-	 */
-	const cssStringToObject = ( val ) => {
-		const tree = [ {} ];
-		let block, left;
-
-		while ( ( block = newRule.exec( val.replace( ruleClean, '' ) ) ) ) {
-			if ( block[ 4 ] ) {
-				tree.shift();
-			} else if ( block[ 3 ] ) {
-				left = block[ 3 ].replace( ruleNewline, empty ).trim();
-				tree.unshift( ( tree[ 0 ][ left ] = tree[ 0 ][ left ] || {} ) );
-			} else {
-				tree[ 0 ][ block[ 1 ] ] = block[ 2 ]
-					.replace( ruleNewline, empty )
-					.trim();
-			}
-		}
-
-		return tree[ 0 ];
-	};
 
 	// data-wp-style--[style-key]
 	directive( 'style', ( { directives: { style }, element, evaluate } ) => {
@@ -182,7 +205,7 @@ export default () => {
 				if ( ! result ) delete element.props.style[ key ];
 				else element.props.style[ key ] = result;
 
-				useEffect( () => {
+				useInit( () => {
 					// This seems necessary because Preact doesn't change the styles on
 					// the hydration, so we have to do it manually. It doesn't need deps
 					// because it only needs to do it the first time.
@@ -191,7 +214,7 @@ export default () => {
 					} else {
 						element.ref.current.style[ key ] = result;
 					}
-				}, [] );
+				} );
 			} );
 	} );
 
@@ -202,22 +225,11 @@ export default () => {
 				const attribute = entry.suffix;
 				const result = evaluate( entry );
 				element.props[ attribute ] = result;
-				// Preact doesn't handle the `role` attribute properly, as it doesn't remove it when `null`.
-				// We need this workaround until the following issue is solved:
-				// https://github.com/preactjs/preact/issues/4136
-				useLayoutEffect( () => {
-					if (
-						attribute === 'role' &&
-						( result === null || result === undefined )
-					) {
-						element.ref.current.removeAttribute( attribute );
-					}
-				}, [ attribute, result ] );
 
 				// This seems necessary because Preact doesn't change the attributes
 				// on the hydration, so we have to do it manually. It doesn't need
 				// deps because it only needs to do it the first time.
-				useEffect( () => {
+				useInit( () => {
 					const el = element.ref.current;
 
 					// We set the value directly to the corresponding
@@ -260,52 +272,10 @@ export default () => {
 					} else {
 						el.removeAttribute( attribute );
 					}
-				}, [] );
+				} );
 			}
 		);
 	} );
-
-	// data-wp-navigation-link
-	directive(
-		'navigation-link',
-		( {
-			directives: { 'navigation-link': navigationLink },
-			props: { href },
-			element,
-		} ) => {
-			const { value: link } = navigationLink.find(
-				( { suffix } ) => suffix === 'default'
-			);
-
-			useEffect( () => {
-				// Prefetch the page if it is in the directive options.
-				if ( link?.prefetch ) {
-					// prefetch( href );
-				}
-			} );
-
-			// Don't do anything if it's falsy.
-			if ( link !== false ) {
-				element.props.onclick = async ( event ) => {
-					event.preventDefault();
-
-					// Fetch the page (or return it from cache).
-					await navigate( href );
-
-					// Update the scroll, depending on the option. True by default.
-					if ( link?.scroll === 'smooth' ) {
-						window.scrollTo( {
-							top: 0,
-							left: 0,
-							behavior: 'smooth',
-						} );
-					} else if ( link?.scroll !== false ) {
-						window.scrollTo( 0, 0 );
-					}
-				};
-			}
-		}
-	);
 
 	// data-wp-ignore
 	directive(
@@ -330,64 +300,62 @@ export default () => {
 	// data-wp-text
 	directive( 'text', ( { directives: { text }, element, evaluate } ) => {
 		const entry = text.find( ( { suffix } ) => suffix === 'default' );
-		element.props.children = evaluate( entry );
+		try {
+			const result = evaluate( entry );
+			element.props.children =
+				typeof result === 'object' ? null : result.toString();
+		} catch ( e ) {
+			element.props.children = null;
+		}
 	} );
 
-	// data-wp-slot
-	directive(
-		'slot',
-		( { directives: { slot }, props: { children }, element } ) => {
-			const { value } = slot.find(
-				( { suffix } ) => suffix === 'default'
-			);
-			const name = typeof value === 'string' ? value : value.name;
-			const position = value.position || 'children';
+	// data-wp-run
+	directive( 'run', ( { directives: { run }, evaluate } ) => {
+		run.forEach( ( entry ) => evaluate( entry ) );
+	} );
 
-			if ( position === 'before' ) {
+	// data-wp-each--[item]
+	directive(
+		'each',
+		( {
+			directives: { each, 'each-key': eachKey },
+			context: inheritedContext,
+			element,
+			evaluate,
+		} ) => {
+			if ( element.type !== 'template' ) return;
+
+			const { Provider } = inheritedContext;
+			const inheritedValue = useContext( inheritedContext );
+
+			const [ entry ] = each;
+			const { namespace, suffix } = entry;
+
+			const list = evaluate( entry );
+			return list.map( ( item ) => {
+				const mergedContext = deepSignal( {} );
+
+				const itemProp = suffix === 'default' ? 'item' : suffix;
+				const newValue = deepSignal( {
+					[ namespace ]: { [ itemProp ]: item },
+				} );
+				mergeDeepSignals( newValue, inheritedValue );
+				mergeDeepSignals( mergedContext, newValue, true );
+
+				const scope = { ...getScope(), context: mergedContext };
+				const key = eachKey
+					? getEvaluate( { scope } )( eachKey[ 0 ] )
+					: item;
+
 				return (
-					<>
-						<Slot name={ name } />
-						{ children }
-					</>
+					<Provider value={ mergedContext } key={ key }>
+						{ element.props.content }
+					</Provider>
 				);
-			}
-			if ( position === 'after' ) {
-				return (
-					<>
-						{ children }
-						<Slot name={ name } />
-					</>
-				);
-			}
-			if ( position === 'replace' ) {
-				return <Slot name={ name }>{ children }</Slot>;
-			}
-			if ( position === 'children' ) {
-				element.props.children = (
-					<Slot name={ name }>{ element.props.children }</Slot>
-				);
-			}
+			} );
 		},
-		{ priority: 4 }
+		{ priority: 20 }
 	);
 
-	// data-wp-fill
-	directive(
-		'fill',
-		( { directives: { fill }, props: { children }, evaluate } ) => {
-			const entry = fill.find( ( { suffix } ) => suffix === 'default' );
-			const slot = evaluate( entry );
-			return <Fill slot={ slot }>{ children }</Fill>;
-		},
-		{ priority: 4 }
-	);
-
-	// data-wp-slot-provider
-	directive(
-		'slot-provider',
-		( { props: { children } } ) => (
-			<SlotProvider>{ children }</SlotProvider>
-		),
-		{ priority: 4 }
-	);
+	directive( 'each-child', () => null );
 };