+++ packages/compose/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Deprecations
+
+-   The `pure` HoC has been deprecated. Use `memo` or `PureComponent` instead ([#57173](https://github.com/WordPress/gutenberg/pull/57173)).
+
+## 6.26.0 (2024-01-10)
+
 ## 6.25.0 (2023-12-13)
 
 ## 6.24.0 (2023-11-29)